+++ src/main/java/com/example/bankservice/repository/BankRepo.java
@@ -1,9 +1,9 @@
 package com.example.bankservice.repository;
 
-import com.example.bankservice.entity.AppUser;
 import com.example.bankservice.entity.BankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import  java.util.Optional;
 
@@ -12,6 +12,6 @@ public interface BankRepo extends JpaRepository<BankAccount, Long> {
 
     Optional<BankAccount> findByAccountNumber(String accountNumber);
 
-    List<BankAccount> findByBalanceGreaterThan(double minBalance);
+    List<BankAccount> findByBalanceGreaterThan(BigDecimal minBalance);
 }
